Reference

Reference

Map plotting functions

OpenStreetMapXPlot.plotmap — Function.

plotmap(nodes::Dict{Int,T},
             bounds::Union{Nothing,OpenStreetMapX.Bounds{T}} = nothing;
             buildings::Union{Nothing,Vector{OpenStreetMapX.Way}} = nothing,
             buildingStyle::Styles=OpenStreetMapXPlot.Style("0x000000", 1, "-"),
             roadways::Union{Nothing,Vector{OpenStreetMapX.Way}} = nothing,
             roadwayStyle::Styles=OpenStreetMapXPlot.Style("0x007CFF", 1.5, "-"),
             walkways::Union{Nothing,Vector{OpenStreetMapX.Way}} = nothing,
             walkwayStyle::Styles=OpenStreetMapXPlot.Style("0x007CFF", 1.5, "-"),
             cycleways::Union{Nothing,Vector{OpenStreetMapX.Way}} = nothing,
             cyclewayStyle::Styles=OpenStreetMapXPlot.Style("0x007CFF", 1.5, "-"),
             features::Union{Nothing,Dict{Int64,Tuple{String,String}}} = nothing,
             featureStyle::Styles=OpenStreetMapXPlot.Style("0xCC0000", 2.5, "."),
             width::Int=600,
             height::Int=600,
             fontsize::Integer=0,
             km::Bool=false, 
			 use_plain_pyplot=false
			 ) where T<:Union{OpenStreetMapX.LLA,OpenStreetMapX.ENU}

Plots selected map features for a given dictionary of node locations (nodes) and within the given bounds.

The default plotting backend is Plots.jl, however if the use_plain_pyplot is set to true than PyPlot.jl is used (note that due to PyPlot's performance this setting should be used for maps up to few thousands nodes).

Returns an object that can be used for further plot updates.

source

plotmap(m::OpenStreetMapX.MapData;
    roadwayStyle = OpenStreetMapXPlot.LAYER_STANDARD, 
    width=600, height=600, use_plain_pyplot=false)

Plots roadways for a given map m.

The width will be set to width and the height will be set to height.

The default plotting backend is Plots.jl, however if the use_plain_pyplot is set to true than PyPlot.jl is used (note that due to PyPlot's performance this setting should be used for maps up to few thousands nodes).

Returns an object that can be used for further plot updates.

source

OpenStreetMapXPlot.addroute! — Function.

addroute!(p, nodes::Dict{Int,T},
               route::Vector{Int}; 
               route_color::String ="0x000053",
               km::Bool=false, 
               start_name="A", end_name="B", 
               fontsize=15
               ) where T<:Union{OpenStreetMapX.LLA,OpenStreetMapX.ENU}

Adds a route to the plot p using the node information stored in nodes.

The first element from the list of nodes route will be annoted by start_name while the last will be annotated by end_name.

Returns an object that can be used for further plot updates.

source

OpenStreetMapXPlot.plot_nodes! — Function.

plot_nodes!(p, m::OpenStreetMapX.MapData, 
    nodeids::Vector{Int};
    start_numbering_from::Union{Int,Nothing}=1, 
    km::Bool=false, 
    color="darkgreen", 
    fontsize=10
    ) where T<:Union{OpenStreetMapX.LLA,OpenStreetMapX.ENU}

Plots nodes having node identifiers nodeids on the plot p using map information m. By default the node indices within the are plotted (e.g. 1, 2, 3...), however, setting the parameter start_numbering_from to nothing will show actual OSM node identifiers.

Returns an object that can be used for further plot updates.

source